keep the deb files around
the varnish-dev package is needed to build modules, which we want to be
easy to do. However, it would cost us a lot of space to keep it around
due to its dependencies.

So we stash the package files in `/pkgs`for a minimal amount of space,
and people can install it it the need it.